+++ src/test/java/ch/bfh/unicrypt/helper/sequence/SequenceIteratorTest.java
@@ -70,11 +70,11 @@ public class SequenceIteratorTest {
 	public void testSkip() {
 		SequenceIterator<Integer> iterator = SequenceIterator.getInstance(values.iterator());
 		iterator.skip(2);
-		assertEquals(3, (int) iterator.next());
+		assertEquals(3, (long) iterator.next());
 		iterator.skip(0);
-		assertEquals(4, (int) iterator.next());
+		assertEquals(4, (long) iterator.next());
 		iterator.skip(4);
-		assertEquals(9, (int) iterator.next());
+		assertEquals(9, (long) iterator.next());
 		iterator.skip(4);
 		assertFalse(iterator.hasNext());
 	}
@@ -83,27 +83,27 @@ public class SequenceIteratorTest {
 	public void testFind() {
 		SequenceIterator<Integer> iterator = SequenceIterator.getInstance(values.iterator());
 		Predicate<Integer> pred = value -> value % 2 == 0;
-		assertEquals(2, (int) iterator.find(pred));
-		assertEquals(4, (int) iterator.find(pred));
-		assertEquals(6, (int) iterator.find(pred));
-		assertEquals(8, (int) iterator.find(pred));
-		assertEquals(10, (int) iterator.find(pred));
+		assertEquals(2, (long) iterator.find(pred));
+		assertEquals(4, (long) iterator.find(pred));
+		assertEquals(6, (long) iterator.find(pred));
+		assertEquals(8, (long) iterator.find(pred));
+		assertEquals(10, (long) iterator.find(pred));
 		assertEquals(null, iterator.find(pred));
 	}
 
 	@Test
 	public void testNext() {
 		SequenceIterator<Integer> iterator = SequenceIterator.getInstance(values.iterator());
-		assertEquals(1, (int) iterator.next());
-		assertEquals(2, (int) iterator.next());
-		assertEquals(3, (int) iterator.next());
-		assertEquals(4, (int) iterator.next());
-		assertEquals(5, (int) iterator.next());
-		assertEquals(6, (int) iterator.next());
-		assertEquals(7, (int) iterator.next());
-		assertEquals(8, (int) iterator.next());
-		assertEquals(9, (int) iterator.next());
-		assertEquals(10, (int) iterator.next());
+		assertEquals(1, (long) iterator.next());
+		assertEquals(2, (long) iterator.next());
+		assertEquals(3, (long) iterator.next());
+		assertEquals(4, (long) iterator.next());
+		assertEquals(5, (long) iterator.next());
+		assertEquals(6, (long) iterator.next());
+		assertEquals(7, (long) iterator.next());
+		assertEquals(8, (long) iterator.next());
+		assertEquals(9, (long) iterator.next());
+		assertEquals(10, (long) iterator.next());
 		assertFalse(iterator.hasNext());
 	}
